feat: add PriceChart component for visualizing price data in TickerMetrics
